cleanup all the code, next stop: getting it working
