+++ src/OmniSharp/Api/Refactoring/OmnisharpController.CodeActions.cs
@@ -1,15 +1,12 @@
-﻿#if ASPNET50
-using ICSharpCode.NRefactory6.CSharp.Refactoring;
-#endif
-using Microsoft.AspNet.Mvc;
+﻿using Microsoft.AspNet.Mvc;
 using Microsoft.CodeAnalysis.CodeActions;
 using Microsoft.CodeAnalysis.CodeRefactorings;
 using Microsoft.CodeAnalysis.Text;
 using OmniSharp.Models;
+using OmniSharp.Services;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Reflection;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -28,40 +25,37 @@ namespace OmniSharp
         }
 
         [HttpPost("runcodeaction")]
-        public async Task<RunCodeActionsResponse> RunCodeAction([FromBody]CodeActionRequest request)
+        public async Task<RunCodeActionResponse> RunCodeAction([FromBody]CodeActionRequest request)
         {
             var actions = new List<CodeAction>();
             var context = await GetContext(request, actions);
             await GetContextualCodeActions(request,context);
             if (request.CodeAction > actions.Count())
-                return new RunCodeActionsResponse();
+                return new RunCodeActionResponse();
 
-            //run the code action
-
+            var action = actions.ElementAt(request.CodeAction);
+            //this line fails \/ ;(
+            var preview = await action.GetPreviewOperationsAsync(CancellationToken.None);
 
             //return the new document
             var sourceText = await context.Value.Document.GetTextAsync();
-            return new RunCodeActionsResponse { Text = sourceText.ToString() };
+            return new RunCodeActionResponse { Text = sourceText.ToString() };
         }
-
 
         private async Task<CodeRefactoringContext?> GetContext(CodeActionRequest request, List<CodeAction> actionsDestination)
         {
             var document = _workspace.GetDocument(request.FileName);
-
             if (document != null)
             {
                 var sourceText = await document.GetTextAsync();
                 var position = sourceText.Lines.GetPosition(new LinePosition(request.SelectionStartLine.Value - 1, request.SelectionStartColumn.Value - 1));
                 var location = new TextSpan(position, 1);
-
                 return new CodeRefactoringContext(document, location, (a) => actionsDestination.Add(a), new System.Threading.CancellationToken());
             }
             //todo, handle context creation issues
             return null;
         }
-
-        private async Task<bool> GetContextualCodeActions(CodeActionRequest request, CodeRefactoringContext? context)
+        private async Task GetContextualCodeActions(CodeActionRequest request, CodeRefactoringContext? context)
         {
             var providers = new CodeActionProviders().GetProviders();
             if (context.HasValue)
@@ -74,49 +68,8 @@ namespace OmniSharp
                     }
                     catch (Exception) { }
                 }
-                return true;
             }
-            return false;
         }
-
-    }
-
-
-    public class CodeActionRequest : Request
-    {
-        public int CodeAction { get; set; }
-        public int? SelectionStartColumn { get; set; }
-        public int? SelectionStartLine { get; set; }
-        public int? SelectionEndColumn { get; set; }
-        public int? SelectionEndLine { get; set; }
-    }
-
-    public class CodeActionProviders
-    {
-        public IEnumerable<CodeRefactoringProvider> GetProviders()
-        {
-            var types = Assembly.GetAssembly(typeof(UseVarKeywordAction))
-                                .GetTypes()
-                                .Where(t => typeof(CodeRefactoringProvider).IsAssignableFrom(t));
-
-            IEnumerable<CodeRefactoringProvider> providers =
-                types
-                    .Where(type => !type.IsInterface
-                            && !type.IsAbstract
-                            && !type.ContainsGenericParameters) //TODO: handle providers with generic params 
-                    .Select(type => (CodeRefactoringProvider)Activator.CreateInstance(type));
-
-            return providers;
-        }
-    }
-
-    public class GetCodeActionsResponse
-    {
-        public IEnumerable<string> CodeActions { get; set; }
-    }
-    public class RunCodeActionsResponse
-    {
-        public string Text { get; set; }
     }
 #endif
 }
