+++ src/OmniSharp/Models/GetCodeActionsResponse.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace OmniSharp.Models
+{
+    public class GetCodeActionsResponse
+    {
+        public IEnumerable<string> CodeActions { get; set; }
+    }
+}
