+++ src/OmniSharp/Models/RunCodeActionResponse.cs
@@ -1,0 +1,7 @@
+﻿namespace OmniSharp.Models
+{
+    public class RunCodeActionResponse
+    {
+        public string Text { get; set; }
+    }
+}
