+++ src/OmniSharp/project.json
@@ -3,7 +3,8 @@
     "dependencies": {
         "Kestrel": "1.0.0-beta2",
         "Microsoft.AspNet.Mvc": "6.0.0-beta2",
-        "Microsoft.AspNet.Diagnostics": "1.0.0-beta2",
+		"Microsoft.AspNet.Diagnostics": "1.0.0-beta2",
+        "Microsoft.CodeAnalysis": "1.0.0-beta2",
         "Microsoft.CodeAnalysis.CSharp": "1.0.0-beta2",
         "Microsoft.CodeAnalysis.CSharp.Workspaces": "1.0.0-beta2",
         "Microsoft.CodeAnalysis.VisualBasic.Workspaces": "1.0.0-beta2",
